+++ nixarr/bazarr/default.nix
@@ -66,6 +66,18 @@ in {
         Route Bazarr traffic through the VPN.
       '';
     };
+
+    vpn.configureNginx = mkOption {
+      type = types.bool;
+      default = cfg.vpn.enable;
+      example = false;
+      description = ''
+        **Required options:** [`nixarr.bazarr.vpn.enable`)(#nixarr.bazarr.vpn.enable)
+
+        Configure nginx as a reverse proxy for the Bazarr web ui.
+      '';
+      defaultText = literalExpression "nixarr.bazarr.vpn.enable";
+    };
   };
 
   config = mkIf (nixarr.enable && cfg.enable) {
@@ -75,6 +87,13 @@ in {
         message = ''
           The nixarr.bazarr.vpn.enable option requires the
           nixarr.vpn.enable option to be set, but it was not.
+        '';
+      }
+      {
+        assertion = cfg.vpn.configureNginx -> cfg.vpn.enable;
+        message = ''
+          The nixarr.bazarr.vpn.configureNginx option requires the
+          nixarr.bazarr.vpn.enable option to be set, but it was not.
         '';
       }
     ];
@@ -92,6 +111,9 @@ in {
         Type = "simple";
         User = globals.bazarr.user;
         Group = globals.bazarr.group;
+        # Set UMask to 0002 so directories are created with group write permission (775)
+        # This allows other services in the media group (like Jellyfin) to modify files
+        UMask = "0002";
         SyslogIdentifier = "bazarr";
         ExecStart = pkgs.writeShellScript "start-bazarr" ''
           ${pkgs.bazarr}/bin/bazarr \
@@ -134,7 +156,7 @@ in {
       ];
     };
 
-    services.nginx = mkIf cfg.vpn.enable {
+    services.nginx = mkIf cfg.vpn.configureNginx {
       enable = true;
 
       recommendedTlsSettings = true;