+++ nixarr/jellyseerr/default.nix
@@ -68,6 +68,18 @@ in {
       '';
     };
 
+    vpn.configureNginx = mkOption {
+      type = types.bool;
+      default = cfg.vpn.enable;
+      example = false;
+      description = ''
+        **Required options:** [`nixarr.jellyseerr.vpn.enable`)(#nixarr.jellyseerr.vpn.enable)
+
+        Configure nginx as a reverse proxy for the Jellyseerr web ui.
+      '';
+      defaultText = literalExpression "nixarr.jellyseerr.vpn.enable";
+    };
+
     expose = {
       https = {
         enable = mkOption {
@@ -123,6 +135,13 @@ in {
         message = ''
           The nixarr.jellyseerr.vpn.enable option conflicts with the
           nixarr.jellyseerr.expose.https.enable option. You cannot set both.
+        '';
+      }
+      {
+        assertion = cfg.vpn.configureNginx -> cfg.vpn.enable;
+        message = ''
+          The nixarr.jellyseerr.vpn.configureNginx option requires the
+          nixarr.jellyseerr.vpn.enable option to be set, but it was not.
         '';
       }
       {
@@ -208,7 +227,7 @@ in {
     };
 
     services.nginx = mkMerge [
-      (mkIf (cfg.expose.https.enable || cfg.vpn.enable) {
+      (mkIf (cfg.expose.https.enable || cfg.vpn.configureNginx) {
         enable = true;
 
         recommendedTlsSettings = true;
@@ -226,7 +245,7 @@ in {
           };
         };
       })
-      (mkIf cfg.vpn.enable {
+      (mkIf cfg.vpn.configureNginx {
         virtualHosts."127.0.0.1:${builtins.toString cfg.port}" = {
           listen = [
             {