+++ nixarr/radarr/default.nix
@@ -66,6 +66,18 @@ in {
         Route Radarr traffic through the VPN.
       '';
     };
+
+    vpn.configureNginx = mkOption {
+      type = types.bool;
+      default = cfg.vpn.enable;
+      example = false;
+      description = ''
+        **Required options:** [`nixarr.radarr.vpn.enable`)(#nixarr.radarr.vpn.enable)
+
+        Configure nginx as a reverse proxy for the Radarr web ui.
+      '';
+      defaultText = literalExpression "nixarr.radarr.vpn.enable";
+    };
   };
 
   config = mkIf (nixarr.enable && cfg.enable) {
@@ -75,6 +87,13 @@ in {
         message = ''
           The nixarr.radarr.vpn.enable option requires the
           nixarr.vpn.enable option to be set, but it was not.
+        '';
+      }
+      {
+        assertion = cfg.vpn.configureNginx -> cfg.vpn.enable;
+        message = ''
+          The nixarr.radarr.vpn.configureNginx option requires the
+          nixarr.radarr.vpn.enable option to be set, but it was not.
         '';
       }
     ];
@@ -103,6 +122,10 @@ in {
       dataDir = cfg.stateDir;
     };
 
+    # Set UMask to 0002 so directories are created with group write permission (775)
+    # This allows other services in the media group (like Jellyfin) to modify files
+    systemd.services.radarr.serviceConfig.UMask = "0002";
+
     # Enable and specify VPN namespace to confine service in.
     systemd.services.radarr.vpnConfinement = mkIf cfg.vpn.enable {
       enable = true;
@@ -119,7 +142,7 @@ in {
       ];
     };
 
-    services.nginx = mkIf cfg.vpn.enable {
+    services.nginx = mkIf cfg.vpn.configureNginx {
       enable = true;
 
       recommendedTlsSettings = true;